+++ src/test/ExSln2/LedBlinker/mem_test/XyzPointI8.cs
@@ -10,8 +10,8 @@ public class XyzPointI8 : FinObj
     public i8 x = -1, y = 1;
     public i8 z;
 
-    public XyzPointI8()
+    public XyzPointI8(i8 z)
     {
-
+        this.z = z;
     }
 }