+++ src/test/ExSln2/LedBlinker/mem_test/XyzPointUser.cs
@@ -1,0 +1,15 @@
+﻿using finlang;
+
+namespace ex_mem;
+
+/// <summary>
+/// Example for testing constructor with argument
+/// </summary>
+public class XyzPointUser : FinObj
+{
+    [mem] XyzPointI8 p1 = mem.init(new XyzPointI8(55));
+
+    public XyzPointUser()
+    {
+    }
+}