+++ src/test/ExSln2/c99/gen/ex_mem_DoublePoint.c
@@ -11,7 +11,7 @@
 void ex_mem_DoublePoint_ctor(ex_mem_DoublePoint * self)
 {
     memset(self, 0, sizeof(*self));
-    self->start = finlang_mem_init(new ex_mem_XyPointU8 *());
+    ex_mem_XyPointU8_ctor(&self->start);
     self->end = finlang_mem_init(new ex_mem_XyPointU8 *());
 }